Fix symbol search lure issue
Previously, we restricted the lure symbol to occupy only the top position of a
bottom card. We did this to follow a previous implementation peculiarity. Now,
the lure symbol can occupy either the top or bottom position. Or, using the
lure_position_on_card parameter, it can be fixed to a desired position in all trials.