Add module option to new schematic
To aid the creation of m2c2kit assessment modules (assessments that will
be reused in other projects), the new schematic has a --module option.
This will scaffold a folder that will build a module, but it will also
create a simple runner for testing/running. The new schematic will
also, by default, create a git repo with an initial commit unless the
option --skipGit is added.