Update Timer; create LegacyTimer
Previous timer functionality had a weakness such that Timer.start()
would create a new timer OR start a stopped timer. This could lead to
errors if a timer was stopped at the end of a trial, not removed, and
then started again with a new trial. Rather than begin anew from 0,
the timer would continue from its previous elapsed time. To address
this, Timer has been changed so that Timer.start() can only start
an existing timer in a stopped state. Timer.startNew() will
unambiguously create a new timer and start it from 0. Thus, to use the
changed Timer class, old code  using Timer.start() can in most cases now
call Timer.startNew(). Timer.startNew() is idempotent, so it can be
called at the beginning of a trial whether there is an existing timer
with the same name or not. Going forward, all code should use this
changed Timer class, but if the old behavior is desired (not
recommended!), the LegacyTimer class can be used.