Don't pin CLI generated package.json session dependency
In the package.json for a m2c2kit project generated by the CLI, do not
pin the session package version. This is because the goal is for session
to be updated yet still work with assessments older versions of core and
addons. Session is just a runner and should work with different core
addons versions.