Fix copyAssets in build-helpers
the copyAssets rollup plugin discovered assessment id by parsing the
bundle js and finding the id property in the GameOptions object. This
was difficult and brittle. Now, for modules, the assessment id will be
stored in package.json in the m2c2kit.assessmentId property. When using
copyAssets, the id of the assessment will be provided in the id option
if the assessment is being developed locally in this package. OR, the
assessment package name will be provided in the package option. It is
possible for an m2c2kit app to use both a locally developed assessment
and one in an external package, so both id and package can be provided
in CopyAssetsOptions.