Support sequence and group with play Action
Playing and decoding sounds in the browser first require a user
interaction. This complicates actions because the duration of the sound
is not known immediately. This commit adds a class called Futurable that
allow us to assign a duration that may not currently be known but could
be known in the future. The Action class now internally treats some
durations as a Futurable, rather than a number. This commit also has
some refactoring of the Action class.