Add Repeat and RepeatForever actions
To add Repeat, RepeatForever actions, substantial refactoring of the
Action code was needed. Starting out, Actions were always assumed to
proceed in a linear fashion, and once initialized, all the actions were
flattened, even if they had parent Sequence or Group actions (this is
because once the action was initialized, the Sequence and Group
information was no longer needed). However, to support repeating
actions, the Sequence and Group structure needed to be retained. The
refactored code keeps the parent action structure and benefits from more
use of recursion when evaluating and calculating actions.