Add instructions game parameter
There is now an instructions game parameter that can be used to
override all the built-in instructions for a game. This is useful when
there is a need for heavy customization of instructions (e.g., adding
new scenes) instead of just changing a string. This has been added to
color dots, color shapes, grid memory, and symbol search.
The de-DE symbol search instructions have also been updated.