Generate schemas when building assessments
When building an assessment, the package.json build script will now
extract all game schemas (GameParameters, TrialSchema) with the
m2c2kit schema-util and place them in schemas.json in the root of
the assessment. This file will be included with artifacts distributed to
npm. This is so other consumers of the assessment can access the
schemas without having to execute the assessment. Other consumers
might be a data analysis program, a codebook creation program, or
an assessment runner.