Get automaticTrialSchema from core bundle
Previously, to get the automaticTrialSchema used in a game, the
schema-util needed access to the Game.ts source file. This is
inconvenient if a game is not being developed in the main
m2c2kit monorepo. Now, if the --game-class option is omitted,
schema-util will search for node_modules/@m2c2kit/core/dist/index.js
and parse the bundle for the automaticTrialSchema.