Add runners to assessments
For assessments in the monorepo, add a simple runner script to each
assessment's package so that the assessment can be easily run for
development and debugging. This required adding a separate
tsconfig and rollup config for building the assessment in "debug"
mode using the runner. It also meant that @m2c2kit/session is now
a dependency for each assessment. However, this should be considered
a dev dependency, since it @m2c2kit/session is used only for the
debug/testing build.