Update versions, CHANGELOG, Game.setParameters()
Game.setParameters() will try to convert a value to the target type,
and it throws an error if the conversion fails. For object and array
targets, the string value can be the stringified JSON representation
of the object or array or a uri-encoded string of the JSON
representation. A warning will show if an array or object is converted
to a string.