Suppress "nothing" CLI message, add --list to static-site
If the CLI does not modify the file system, do not print a "Nothing
to be done" message. The static-site schematic now has a --list
option that will download an asset from @m2c2kit/assessments-registry
package that contains the current m2c2kit assessments on NPM and
display these.