Don't set DOM events as handled when dragging
Previously, if the user was actively dragging a node, the DOM pointer
move and pointer up events would be processed by the node's dragging
logic, but then marked as handled so that these DOM events would not
propagate to other nodes. The reasoning was that if multiple draggable
nodes were activated, more than one could be dragged, thus the first
node to process the DOM event should swallow it. However, it now seems
this is the wrong choice. For example, a node may be dragging, and
another node may be listening to a pointer move event. With the changes
in this commit, DOM events on dragging nodes will be allowed to be
passed on and processed by other nodes.