Fix DOM events on child nodes with tied zPosition
DOM pointer events (pointer up, move, leave, down) are processed on
child nodes according to zPosition. If two nodes are overlapping, the
child node drawn on top (higher zPosition) should receive the
pointer event first. However, if two child nodes had the same zPosition,
the earlier child in the array (which would have been drawn "under" the
late child in the array) would receive the pointer event first. This is
now fixed so that among child nodes with the same zPosition, the child
that is visually "on top" (the one with the later position in the child
array) receives the pointer event first.