Improve integration test scripts and dockerfiles
The integration test requires m2c2kit libraries to be published to the
temporary registry (verdaccio). Previously, containers that needed these
libraries would continually poll the registry to see if the libraries
were ready. This was fragile, and sometimes libraries were not available
even if polling said they were. In this new approach, the containers
that need the libraries listen for a TCP message from the publish
container that the libraries are ready. netcat is used to send and
listen for messages.