Fix Label underline; add strikethrough,overline
Previously, underlining within label text was done by manually drawing a
line beneath the text segment. This caused some visual inconsistencies
on desktop versus mobile browser. Now, the underline decoration native
to canvaskit is applied, and the canvaskit library takes care of it.
Strikethrough and overline decoration support is also added; tags for
this are 's' and 'o'.