Update data-calc
Allow summarize operations to accept arguments for values and arrays of
values, not just a variable name in the current DataCalc object. This
allows more flexibility of data recoding. Also: the coerceBooleans
option defaults to true. When passing only the start argument to the
slice method, it returns a single observation at the start index. Slice
also support negative indexes. Passing into the pull method a variable
name that is not present in the DataCalc object now returns null, not an
error.