Support scoring calculations in games
Games can now calculate their own scores from their assessment data. To
do this, they implement the ScoringProvider interface, which is the
calculateScores() method. In addition, scoring schema must be specified
in the game options. calculateScores() simply generates the scores. To
propagate them back to the host session, addScoringData() adds them
to the game, and scoringComplete() marks the scoring as complete and
triggers a callback to send the scoring data to the session.