Add scoring to symbol search
Once a participant completes all trials, symbol search now generates
scores that summarize the trials. Scores are generated only if the game
parameter "scoring" is set to true. Currently, the default of "scoring"
is false. Once we have tested and developed scoring further, the
default for scoring will be true. Also: update the assessments demo
to indicate scoring data in the console.