Use custom M2Error class for errors
When errors are thrown within the m2c2kit framework, throw an error of
type M2Error, rather than Error, so we can identifying it as coming
from the m2c2kit framework. M2Error extends the standard Error class,
so it behaves the same as Error, and the only difference is that its
name property is M2Error.